fix: Go scalePositionToFCL 用浮点除法匹配 Python 的四舍五入
Go 的 scalePositionToFCL 之前用整数除法截断 (clampInt(value)/10)，
而 Python 的 clamp_int(clamp_int(value) / 10) 是浮点除法 + round()。
ZL 位置 1236 在 Python 中正确舍入到 FCL 124，但 Go 截断到 123。

改用 float64 除法后，两个实现的行为一致。
该函数当前仅在 ZL→FCL 路径中使用（Go JNI 库暂未实现 zl2fcl），
但修正移植确保将来复用时不会产生偏差。

+++ go/utils.go
@@ -244,8 +244,10 @@ func minInt(a, b int) int {
 }
 
 // scalePositionToFCL: FCL uses 0..1000, ZL uses 0..10000.
+// Python uses clamp_int(clamp_int(value) / 10) which is float division + round,
+// so we must use float division here too (integer division would truncate).
 func scalePositionToFCL(value interface{}) int {
-	return maxInt(0, minInt(1000, clampInt(clampInt(value)/10)))
+	return maxInt(0, minInt(1000, clampInt(float64(clampInt(value))/10)))
 }
 
 // scalePositionToZL: FCL uses 0..1000, ZL uses 0..10000.
